Move more tests around
